Fix bug about reference counting index nodes.
When inserting, if we do not need to create a new root node,
we do anyway, to keep track of root node id. That's a little
bit wasteful, but OK. However, previously we did not decrement
the refcount for the temporary index node, which meant it did
not get freed, which meant the leaf nodes accessible via it
did not get freed, resulting in a lot of wasted space and time.

"./speedtest testdir 10000 no" went from 575/s to 1259/s.