Fix NodeStoreMemory.put_node to freeze nodes.
We do not need to make the same change in get_node,
since the only way to have nodes in an in-memory node
store is by using put_node. So by definition, all nodes
in the store are frozen.

If someone goes behind our back to unfreeze a node,
that's SEP.