Several changes that needed to be made together during debugging.
Remove assertion of new node's refcount in set_root.
We occasionally need to pass in a new root with a refcount of 2,
specifically when the new root is a child of the old root.
Otherwise the new root gets removed when the old root gets removed.

Replace self.root.keys()[0] with self.root.first_key().
Clearer, and potentially a teeny bit faster, too.

Fix _reduce_height to not modify the about-to-be-removed root in place.

Add tests for _reduce_height to verify that it works.