Remove push_upload_queue from NodeStore API, add commit
Also, re-arrange things so that commit does things in the
right order in order to get everything to disk before the
new metadata file gets written. It's important to write the
metadata last, so that in case of a crash, commit will
continue the next time the journal is opened.