Fix KeyError for missing refcount
We would sometimes, rarely, have a crash that looked like this:

      File "..refcountstore.py", line 76, in get_refcount
        return self.refcounts[node_id]
    KeyError: 32970

In other words, the object keeping track of reference counts for
B-tree nodes would think it doesn't have the reference count for a
node, when it should have it.

After much debugging, by myself and Itamar Turner-Trauring, and
greatly helped by a repeatable test case provided by Rob Kendrick,
this was tracked to an optimisation in the set_refcount method, which
would remove (to save memory) the refcount for a node (remove the
entry for a node's id in the refcount dict), when the refcount was set
to zero. Unfortunately this conflicted with an assumption in
get_refcount that a refcount for a node that was in the dirty set was
actually present in the refcount dict.

The bug fix is easy: don't remove the node from the refcount dict, and
eat the (very minor) memory consumption hit, in the name of
correctness.